Add the Mag Secure Door.
Adds the Mag Secure Door, a Private Secure Door crafted with a Mag Reader to combine the functions.  Only the bottom of the door can be clicked with a Mag Card to produce a magData event.  Also re-added the ability for doors to act as cables so you don't have to run cables if you place another component adjacent to a door.